+++ src/Products/PageTemplates/engine.py
@@ -1,35 +1,40 @@
+"""``chameleon`` integration.
+
+We use the ``zope.tales`` expression implementation (as adapted for ``Zope``)
+not that of ``chemeleon.tales``.
+"""
+
+import ast
+import copy
 import logging
 import re
+import uuid
+import weakref
 
+from chameleon.astutil import Static
+from chameleon.astutil import Symbol
+from chameleon.codegen import template
 from chameleon.tal import RepeatDict
-from chameleon.tales import NotExpr
-from chameleon.tales import StringExpr
 from chameleon.zpt.template import Macros
 
 from AccessControl.class_init import InitializeClass
 from AccessControl.SecurityInfo import ClassSecurityInfo
-from Products.PageTemplates import ZRPythonExpr
+from MultiMapping import MultiMapping
 from Products.PageTemplates.Expressions import getEngine
-from z3c.pt.expressions import ProviderExpr
-from z3c.pt.expressions import PythonExpr
 from z3c.pt.pagetemplate import PageTemplate as ChameleonPageTemplate
 from zope.interface import implementer
 from zope.interface import provider
 from zope.pagetemplate.interfaces import IPageTemplateEngine
 from zope.pagetemplate.interfaces import IPageTemplateProgram
 
-from .expression import ExistsExpr
-from .expression import NocallExpr
-from .expression import PathExpr
-from .expression import TrustedPathExpr
-from .expression import UntrustedPythonExpr
 
-
-# Declare Chameleon's repeat dictionary public
+# Declare Chameleon's repeat dictionary public but prevent modifications
+#  by untrusted code
 RepeatDict.security = ClassSecurityInfo()
 RepeatDict.security.declareObjectPublic()
+RepeatDict.security.declarePrivate(
+    *(set(dir(RepeatDict)) - set(dir(MultiMapping))))
 RepeatDict.__allow_access_to_unprotected_subobjects__ = True
-
 InitializeClass(RepeatDict)
 
 # Declare Chameleon Macros object accessible
@@ -40,38 +45,107 @@ re_match_pi = re.compile(r'<\?python([^\w].*?)\?>', re.DOTALL)
 logger = logging.getLogger('Products.PageTemplates')
 
 
+# zt_expr registry management
+#  ``chameleon`` compiles TALES expressions to Python byte code
+#  which has no support for arbitrary Python expressions but
+#  can handle global (called `Static`) objects which include
+#  globally defined functions.
+#  We must pass ``MappedExpr`` instances to the ``zope`` TALES
+#  expressions. We use a zt_expr registry accessed via the global
+#  function ``_fetch_zt_expr``.
+
+_zt_expr_registry = {}
+
+
+def _fetch_zt_expr(eid):
+    return _zt_expr_registry[eid]
+
+
+_fetch_zt_expr_node = Static(Symbol(_fetch_zt_expr))
+
+
+def _del_zt_expr(ref):
+    try:
+        del _zt_expr_registry[ref.eid]
+    except KeyError:
+        pass
+
+
+def _c_context_2_z_context(c_context):
+    z_context = copy.copy(c_context["__zt_context__"])
+    # not sure that ``lazy`` works as expected
+    z_context.vars = dict(c_context.vars)
+    return z_context
+
+
+_c_context_2_z_context_node = Static(Symbol(_c_context_2_z_context))
+
+
+class LifetimeRef(weakref.ref):
+    atexit = False
+
+    def __init__(self, obj, callback, eid):
+        self.eid = eid
+        super(LifetimeRef, self).__init__(obj, callback)
+
+
+class MappedExpr(object):
+    """map expression: ``zope.tales`` --> ``chameleon.tales``."""
+    def __init__(self, z_expr, lifetime_ref):
+        eid = self.eid = uuid.uuid4().int
+        z_expr.lifetime_ref = \
+            LifetimeRef(lifetime_ref(), _del_zt_expr, eid=eid)
+        _zt_expr_registry[eid] = z_expr
+
+    def __call__(self, target, c_engine):
+        return template(
+            "target = fetch_zt_expr(eid)(c2z_context(econtext))",
+            target=target,
+            fetch_zt_expr=_fetch_zt_expr_node,
+            eid=ast.Num(self.eid),
+            c2z_context=_c_context_2_z_context_node)
+
+
+class MappedExprType(object):
+    """map expression type: ``zope.tales`` --> ``chameleon.tales``."""
+    def __init__(self, type, handler, engine, lifetime_ref):
+        self.type = type
+        self.handler = handler
+        self.engine = engine
+        self.lifetime_ref = lifetime_ref
+
+    def __call__(self, expression):
+        return MappedExpr(self.handler(self.type, expression, self.engine),
+                          self.lifetime_ref)
+
+
+class Lifetime(object):
+    """Auxiliary class to clean up resources when a template is deleted."""
+
+
+class ChPageTemplate(ChameleonPageTemplate):
+    """override to get an integrated default marker."""
+    def _compile(self, body, builtins):
+        code = super(ChPageTemplate, self)._compile(body, builtins)
+        # redefine ``__default`` as ``zope.tales.tales._default``
+        #  Potentially this could be simpler
+        frags = code.split("\n", 5)
+        for i, frag in enumerate(frags[:-1]):
+            if frag.startswith("__default = "):
+                frags[i] = "from zope.tales.tales import _default as __default"
+                break
+        else:
+            raise RuntimeError("unexpected code prelude %s" % frags[:-1])
+        return "\n".join(frags)
+
+
 @implementer(IPageTemplateProgram)
 @provider(IPageTemplateEngine)
 class Program(object):
 
-    # Zope 2 Page Template expressions
-    secure_expression_types = {
-        'python': UntrustedPythonExpr,
-        'string': StringExpr,
-        'not': NotExpr,
-        'exists': ExistsExpr,
-        'path': PathExpr,
-        'provider': ProviderExpr,
-        'nocall': NocallExpr,
-    }
-
-    # Zope 3 Page Template expressions
-    expression_types = {
-        'python': PythonExpr,
-        'string': StringExpr,
-        'not': NotExpr,
-        'exists': ExistsExpr,
-        'path': TrustedPathExpr,
-        'provider': ProviderExpr,
-        'nocall': NocallExpr,
-    }
-
-    extra_builtins = {
-        'modules': ZRPythonExpr._SecureModuleImporter()
-    }
-
-    def __init__(self, template):
+    def __init__(self, template, lifetime):
         self.template = template
+        self.lifetime = lifetime
 
     def __call__(self, context, macros, tal=True, **options):
         if tal is False:
@@ -82,8 +156,8 @@ class Program(object):
         # in turn used by the secure Python expression
         # implementation whenever a 'repeat' symbol is found
         kwargs = context.vars
-        kwargs['wrapped_repeat'] = kwargs['repeat']
         kwargs['repeat'] = RepeatDict(context.repeat_vars)
+        kwargs["__zt_context__"] = context
 
         return self.template.render(**kwargs)
 
@@ -108,9 +182,12 @@ class Program(object):
                     )
                 )
 
-            expression_types = cls.secure_expression_types
-        else:
-            expression_types = cls.expression_types
+        lifetime = Lifetime()
+        lifetime_ref = weakref.ref(lifetime)
+
+        expr_types = {}
+        for ty, expr in engine.types.items():
+            expr_types[ty] = MappedExprType(ty, expr, engine, lifetime_ref)
 
         # BBB: Support CMFCore's FSPagetemplateFile formatting
         if source_file is not None and source_file.startswith('file:'):
@@ -120,10 +197,10 @@ class Program(object):
             # Default to '<string>'
             source_file = ChameleonPageTemplate.filename
 
-        template = ChameleonPageTemplate(
+        template = ChPageTemplate(
             text, filename=source_file, keep_body=True,
-            expression_types=expression_types,
-            encoding='utf-8', extra_builtins=cls.extra_builtins,
+            expression_types=expr_types,
+            encoding='utf-8',
         )
 
-        return cls(template), template.macros
+        return cls(template, lifetime), template.macros