+++ src/Products/PageTemplates/tests/test_pagetemplate.py
@@ -3,6 +3,7 @@ import unittest
 
 from six import PY3
 
+from AccessControl import safe_builtins
 from Products.PageTemplates.PageTemplateFile import PageTemplateFile
 from Testing.ZopeTestCase import ZopeTestCase
 
@@ -20,6 +21,8 @@ class TestPageTemplateFile(ZopeTestCase):
     def _makeOne(self, name):
         return PageTemplateFile(os.path.join(path, name)).__of__(self.app)
 
+    @unittest.skipIf("sorted" not in safe_builtins,
+                     "`sorted` not allowed by `AccessControl`")
     def test_rr(self):
         class Prioritzed(object):
             __allow_access_to_unprotected_subobjects__ = 1