+++ CHANGES.rst
@@ -10,6 +10,10 @@ https://zope.readthedocs.io/en/2.13/CHANGES.html
 4.3.1 (unreleased)
 ------------------
 
+- Fix issue 717 by using ``zope.tales.tales`` expressions even with
+  the ``Chameleon`` template engine.
+  (`#717 <https://github.com/zopefoundation/Zope/issues/797>`_).
+
 - Fixed encoding issue of `displayname` WebDAV property
   (`#797 <https://github.com/zopefoundation/Zope/issues/797>`_)